Refactor a bit and include storage limit
The storage limit is currently calculated using
rootBucket.Stat().LeafInuse, which gives a different result from simply
iterating through all key-values and finding num bytes with len(value).

This is kinda risky considering I don't know exactly what a Leaf is in
this context, so maybe it's a better idea to calculate by iterating
through? Also, might consider storing the bytesUsed in the database
(using a new bucket) instead of a global variable.